pyproject.toml
[project]
name = "gmaps2gpx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "polyline>=2.0.4",
    "python-dotenv>=1.2.2",
]
